requirements.txt
pillow~=8.1.0
pygame~=2.0.1
